Also link to Nexe base image in the header info
It’s important that people can verify the build if they want to and the link to the source for the exact commit as well as the Nexe base image help with that. The link to Nexe source is in the package.json file that’s linked to in the source link so that is not necessary to include.